requirements.txt
# Core dependencies
python-dotenv
scipy

# Speech-to-Text (STT)
transformers
torch
accelerate
soundfile
librosa
faster-whisper

# Text-to-Speech (TTS)
langid
fasttext
gTTS
edge-tts

# Large Language Models (LLM)
langchain-huggingface
langchain-groq
langchain-google-genai

# User Interface
gradio==4.14.0